--- a/build.xml
+++ b/build.xml
@@ -34,6 +34,7 @@
     <property name="classes.test" value="testclasses"/>
     <property name="scripts" value="src/scripts"/>
     <property name="test.output" value="dist/test"/>
+    <property name="intelDeflator" value="${user.dir}/lib/jni/libIntelDeflater.so"/>
 
     <property name="javac.target" value="1.8"/>
     <property name="javac.debug" value="true"/>
@@ -46,14 +47,17 @@
     <property name="testng.verbosity" value="2"/>
     <property name="test.debug.port" value="5005" />  <!-- override on the command line if desired -->
 
-    <condition  property="isUnix">
-        <os family="unix"/>
+    <condition property="isUnix">
+        <and>
+            <os family="unix"/>
+            <not><os family="mac"/></not>
+        </and>
     </condition>
     <target name="set_excluded_test_groups_unix" if="isUnix">
         <property name="excludedTestGroups" value="slow, broken"/>
     </target>
     <target name="set_excluded_test_groups_non_unix" unless="isUnix">
-        <property name="excludedTestGroups" value="slow, broken, unix"/>
+        <property name="excludedTestGroups" value="slow, broken, unix, intel"/>
     </target>
     <target name="set_excluded_test_groups" depends="set_excluded_test_groups_unix,set_excluded_test_groups_non_unix"/>
 
@@ -122,6 +126,8 @@
     <macrodef name="run-test">
         <attribute name="excludedTestGroups" default=""/>
         <attribute name="includedTestGroups" default=""/>
+        <attribute name="additionalJVMarg" default=""/>
+
         <sequential>
                 <taskdef resource="testngtasks" classpathref="classpath"/>
                 <testng suitename="htsjdk-tests" classpathref="classpath" outputdir="${test.output}"
@@ -137,7 +143,7 @@
                         <include name="**/Test*.class"/>
                         <include name="**/*Test.class"/>
                     </classfileset>
-                    <jvmarg value="-Xmx2G"/>
+                    <jvmarg line="-Xmx2G @{additionalJVMarg}"/>
                 </testng>
                 <junitreport todir="${dist}/test" >
                     <fileset dir="${test.output}">
@@ -150,14 +156,20 @@
         </sequential>
     </macrodef>
 
-    <target name="test" depends="compile, set_excluded_test_groups" description="Run unit tests">
-            <run-test excludedTestGroups="${excludedTestGroups}, sra"/>
+
+    <target name="test" depends="compile, set_excluded_test_groups, intel-test" description="Run unit tests">
+            <run-test excludedTestGroups="${excludedTestGroups}, sra, intel"/>
     </target>
 
     <target name="sra-test" depends="compile, set_excluded_test_groups" description="Run SRA unit tests">
         <run-test includedTestGroups="sra" excludedTestGroups="${excludedTestGroups}"/>
     </target>
 
+    <!-- needs to be in a separate target since jvm has different argument -->
+    <target name="intel-test" depends="compile, set_excluded_test_groups" description="Run Intel unit tests" if="isUnix">
+        <echo message="intelDeflator = ${intelDeflator}"/>
+        <run-test includedTestGroups="intel" additionalJVMarg="-Dsamjdk.intel_deflater_so_path=${intelDeflator}"/>
+    </target>
 
     <target name="single-test"
             depends="compile, compile-tests"