@yfarjoun 2016-03-22
while this seems to be passing in travis, the logs show a SIGSEGV:

 
   [testng] #
   [testng] # A fatal error has been detected by the Java Runtime Environment:
   [testng] #
   [testng] #  SIGSEGV (0xb) at pc=0x00007fb39309f854, pid=3330, tid=140409262946048
   [testng] #
   [testng] # JRE version: Java(TM) SE Runtime Environment (8.0_66-b17) (build 1.8.0_66-b17)
   [testng] # Java VM: Java HotSpot(TM) 64-Bit Server VM (25.66-b17 mixed mode linux-amd64 compressed oops)
   [testng] # Problematic frame:
   [testng] # V  [libjvm.so+0x89e854]  Metaspace::allocate(unsigned long, Metaspace::MetadataType)+0x14
   [testng] #
   [testng] # Failed to write core dump. Core dumps have been disabled. To enable core dumping, try "ulimit -c unlimited" before starting Java again
   [testng] #
   [testng] # An error report file with more information is saved as:
   [testng] # /home/travis/build/samtools/htsjdk/hs_err_pid3330.log
   [testng] #
   [testng] # If you would like to submit a bug report, please visit:
   [testng] #   http://bugreport.java.com/bugreport/crash.jsp
   [testng] #