@yfarjoun 2016-03-22
As I mentioned, it fails on our GenuineIntel machine, but succeeds on a AMD machine (both 64-bit)