@mishalinaik 2016-03-23
@yfarjoun Fixed the issue you were facing it was a tricky one. I pushed the commit