@yfarjoun 2016-03-23
@akiezun given that this is a different .so from what you originally tested, could you run your speed-up test on this commit and give it a speed-wise thumbs up? Ideally you would look at the difference between: 

- native jdkDeflator (version 8 is supposed to be faster than before, according to @tfenne)
- previous good version of intelDeflator
- this version of intelDeflator.