@yfarjoun 2016-03-23
@akiezun sorry, I mentioned this somewhere else, but it was probably unclear, could you run your speedup tests on this branch?