@akiezun 2016-03-23
Will do, sorry about the delay

On Wed, Mar 23, 2016 at 4:05 PM, Yossi Farjoun <notifications@github.com>
wrote:

> @akiezun <https://github.com/akiezun> sorry, I mentioned this somewhere
> else, but it was probably unclear, could you run your speedup tests on this
> branch?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/samtools/htsjdk/pull/536#issuecomment-200521774>
>