@yfarjoun 2016-03-24
sure.

On Thu, Mar 24, 2016 at 10:38 AM, Adam Kiezun <notifications@github.com>
wrote:

> @yfarjoun <https://github.com/yfarjoun> do you want level 1 and 5 right?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub
> <https://github.com/samtools/htsjdk/pull/536#issuecomment-200866409>
>