@akiezun 2016-03-24
new code (this branch). 
Linux 2.6.32-573.18.1.el6.x86_64 amd64; Java HotSpot(TM) 64-Bit Server VM 1.8.0-b132
(32GB of heap space), using BAM file of 1.6GB  on NFS.

level 5, JdkDeflater, SyncIO  Elapsed time 212.253 seconds  (baseline)
level 5, IntelDeflater, SyncIO Elapsed time 183.807 seconds  (1.15x  speedup over baseline) 
level 1, JdkDeflater, SyncIO  Elapsed time 135.901 seconds  (1.55x speedup over baseline)
level 1, IntelDeflater, SyncIO Elapsed time 83.993 seconds    (2.55x speedup over baseline)

for fun i looked at asyncIO too (only writing is implemented right now)
level 1, IntelDeflater, AsyncIO Elapsed time 80.150 seconds (2.6x speedup over baseline)