@akiezun 2016-03-24
master (same machine specs etc)
level 5, JdkDeflater, SyncIO  Elapsed time 208.985 seconds
level 5, IntelDeflater, SyncIO Elapsed time 165.721 seconds **<-- this line worried me (new code is slower?)** (2 more reruns of master 169.326, 176.811)
level 1, JdkDeflater, SyncIO  Elapsed time 138.908 seconds
level 1, IntelDeflater, SyncIO Elapsed time 118.816 seconds
level 1, IntelDeflater, AsyncIO Elapsed time 95.011 seconds

So level 1 looks great but level 5 is not so clear cut. To get a better picture, I reran this new branch level 5 IntelDeflater 3 more times:
167.046 seconds
167.476 seconds
168.618 seconds

**Performance Conclusion: performance wise this branch looks good (level 1 improved a lot).**

What I'm unsure about is that file sizes fluctuate:
initial file is 1673819016 bytes
master:
1673819015 mJdk.Level5.sync.bam  <-
1662517443 mIntel.Level5.sync.bam  <- why are those two different sizes (jdk vs intel)?
1899122247 mJdk.Level1.sync.bam   <-
1835785277 mIntel.Level1.sync.bam  <- why are those two different sizes (jdk vs intel)?
1835785277 mIntel.Level1.async.bam (good - same as sync)

new branch:
1673819015 yJdk.Level5.synch.bam <- same as master
1662517443 yIntel.Level5.sync.bam <- same as master
1899122247 yJdk.Level1.sync.bam  <- same as master
2148761044 yIntel.Level1.sync.bam  <- why is size different than jdk and different than master?
2148761044 yIntel.Level1.async.bam (good - same as sync)

@mishalinaik are those changes expected, ie different from JDKDeflater and different between 2 versions of IntelDeflater on level 1?