@mishalinaik 2016-03-24
@akiezun Yes, this is expected. The tradeoff with using igzip to compress at level 1 is you have much higher performance than standard zlib (at compression level 1) but with slightly lower compression ratio (its very close to zlib but not the same). Hope that helps