@akiezun 2016-03-30
@yfarjoun how does the .so file in this PR relate to that that got merged in https://github.com/samtools/htsjdk/pull/539? is it the same, or newer or older (something else)?