@droazen 2016-04-04
@yfarjoun Can this be merged? What's the answer the @akiezun's question above ("how does the .so file in this PR relate to that that got merged in #539? is it the same, or newer or older (something else)?")