@yfarjoun 2016-04-04
This is the only .so version that should be considered to be not broken.