@akiezun 2016-04-05
@jacarey @yfarjoun @droazen are we good to make a 2.0.2 release? it's been 4 months since 2.0.1 - i think this was the last blocking issue