@droazen 2016-04-05
@akiezun Erm, actually it's been exactly 1 month since the last release, which was `2.1.1`. This next release will be `2.2.0`, and yes we are go for launch.